+++ src/ethereum/ethereumInfos.ts
@@ -1,5 +1,6 @@
 import { arbitrum } from './info/arbitrumInfo'
 import { avalanche } from './info/avalancheInfo'
+import { base } from './info/baseInfo'
 import { binancesmartchain } from './info/binancesmartchainInfo'
 import { celo } from './info/celoInfo'
 import { ethDev } from './info/ethDevInfo'
@@ -22,6 +23,7 @@ import { zksync } from './info/zksyncInfo'
 export const ethereumPlugins = {
   arbitrum,
   avalanche,
+  base,
   binancesmartchain,
   celo,
   ethDev,